fix: add seperate variable hooks
